fix: hybrid param-default gate, lazy id-index, MSVC gap doc, module split
Three more Codex review findings (fresh evidence each) plus a structural
split:

1. The pre-v20 Param.default fingerprint-reliability gate only recognized
   a snapshot whose top-level ast_producer was exactly "clang", missing a
   legacy hybrid snapshot's clang-only-appended function (dumper_hybrid.py's
   merge unconditionally stamps that specific function's param_defaults
   provenance as "clang") -- the same false PARAM_DEFAULT_VALUE_CHANGED
   risk, reached through a different producer path. Fixed by having the
   gate consult the caller's already-resolved PER-FUNCTION producer
   (already computed in diff_symbols._diff_param_defaults's own loop)
   instead of the snapshot-level ast_producer, correctly covering pure-clang
   and clang-only-within-hybrid uniformly. Verified all four combinations
   (hybrid pre-v20 artifact suppressed, pure-clang regression check, hybrid
   castxml-backed real change still caught, hybrid both-reliable real
   change still caught).

2. self._id_index() was still built eagerly for a plain LITERAL default
   (e.g. `int timeout = 30;`) even after gating it on hasInClassInitializer/
   _param_has_default -- that gate only skips a field/param with NO
   initializer at all, not one whose value never needs the index in the
   first place. Fixed by threading the index as a genuinely lazy, memoized
   CALLABLE (_IdIndexProvider) instead of a plain dict, invoked only at the
   exact point _canonical_expr resolves a real referencedDecl. Verified via
   a spy that the index is never built for a literal default.

3. An MSVC (clang-cl/--target=*-windows-msvc) specialization member's
   mangled name (e.g. ?VALUE@?$A@H@@2HB) can't be decoded by either
   itanium_scope_components (wrong scheme) or
   diff_cxx_rules.msvc_scope_components (verified it ALSO returns None for
   this exact input, by its own documented design -- a component starting
   with the template marker ?$ is rejected outright, and a specialization's
   class-name component always does). Adding the msvc fallback anyway would
   be dead code on every real input this function ever sees. Documented as
   a known, deliberately-deferred gap (matching AGENTS.md's own established
   reasoning for the sibling MSVC toolchain-profile gap) rather than
   guessed at, with a regression test pinning the current honest behavior
   (falls back to the bare, collision-prone name; never crashes).

Structural: dumper_clang.py crept to 2041 lines (over the 2000-line hard
cap) yet again across this PR's review rounds. Rather than keep shaving
comments, split the whole initializer/default-argument fingerprint chain
(_canonical_expr, _normalize_qual_type, _index_decl_id_qualified_names,
_specialization_scope_key, _expr_fingerprint, _initializer_value,
_init_expr, _unwrap_expr, _field_initializer_value, plus their supporting
constants) into a new leaf module, dumper_clang_expr.py -- mirroring the
diff_symbols_param_defaults.py split earlier in this same PR. Behavior-
preserving: every fingerprint hash verified byte-for-byte identical
before/after across this PR's own real-clang-AST regression fixtures, and
every existing test passed unmodified (the moved names are re-exported back
through dumper_clang's own namespace, so `from abicheck.dumper_clang import
_canonical_expr` etc. still resolves).
