+++ abicheck/diff_symbols_param_defaults.py
@@ -40,31 +40,53 @@ from __future__ import annotations
 from .model import AbiSnapshot
 
 
-def param_default_value_representation_unreliable(snap: AbiSnapshot) -> bool:
-    """True if *snap*'s clang-side non-literal ``Param.default`` fingerprints
-    predate this PR's ``_canonical_expr`` stabilization (schema v20).
+def param_default_value_representation_unreliable(
+    snap: AbiSnapshot, producer: str | None
+) -> bool:
+    """True if *snap*'s non-literal ``Param.default`` fingerprint, for a
+    function whose resolved provenance is *producer*, predates this PR's
+    ``_canonical_expr`` stabilization (schema v20).
 
-    Only meaningful for a FINGERPRINT-shaped value (the ``"expr:"`` prefix)
-    -- a literal default's plain value never touches ``_canonical_expr`` at
-    all, so it stays fully comparable regardless of schema version.
+    *producer* is the caller's already-resolved per-function value (typically
+    ``fact_provenance.fact_producer(snap, func_fact_key(mangled,
+    "param_defaults"))``) — NOT ``snap.ast_producer`` directly. A pure
+    ``"clang"``-producer snapshot and a ``"hybrid"`` snapshot whose merge
+    stamped this SPECIFIC function's ``param_defaults`` provenance as
+    ``"clang"`` (a clang-only-appended function — ``dumper_hybrid.py``'s
+    merge does this unconditionally) share the identical risk (Codex review,
+    fresh evidence, second round): checking ``snap.ast_producer == "clang"``
+    alone missed the hybrid case entirely, since a hybrid snapshot's own
+    top-level producer is ``"hybrid"``, not ``"clang"``, even though the
+    individual function's fingerprint went through the exact same unstable
+    ``_canonical_expr``. Only meaningful for a FINGERPRINT-shaped value (the
+    ``"expr:"`` prefix) -- a literal default's plain value never touches
+    ``_canonical_expr`` at all, so it stays fully comparable regardless of
+    schema version or producer.
     """
     return (
-        snap.from_headers
+        producer == "clang"
+        and snap.from_headers
         and not snap.from_headers_inferred
-        and snap.ast_producer == "clang"
         and not snap.clang_field_initializer_facts_reliable
     )
 
 
 def param_default_fingerprint_comparison_unreliable(
-    old: AbiSnapshot, new: AbiSnapshot, old_value: str, new_value: str
+    old: AbiSnapshot,
+    new: AbiSnapshot,
+    old_producer: str | None,
+    new_producer: str | None,
+    old_value: str,
+    new_value: str,
 ) -> bool:
     """True if comparing *old_value* against *new_value* risks a false
     ``PARAM_DEFAULT_VALUE_CHANGED`` from a fingerprint-algorithm version
-    mismatch rather than a real edit.
+    mismatch rather than a real edit. *old_producer*/*new_producer* are the
+    caller's already-resolved per-function producers — see
+    :func:`param_default_value_representation_unreliable`.
     """
     if not (old_value.startswith("expr:") or new_value.startswith("expr:")):
         return False
     return param_default_value_representation_unreliable(
-        old
-    ) or param_default_value_representation_unreliable(new)
+        old, old_producer
+    ) or param_default_value_representation_unreliable(new, new_producer)
