+++ abicheck/dumper_clang_expr.py
@@ -1,0 +1,396 @@
+# Copyright 2026 Nikolay Petrov
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Direct-clang initializer/default-argument value fingerprinting.
+
+Split out of ``dumper_clang.py`` to stay under its line-count cap — the
+fingerprint chain for ``TypeField.default``/``Param.default`` grew a lot of
+carefully-verified, narrow edge-case handling across several Codex review
+rounds (referenced-declaration identity/scope, template-specialization
+disambiguation, ``sizeof`` operand types, anonymous-type/lambda location
+normalization) and kept pushing the parent file over the hard cap. A leaf
+module (must not import from ``dumper_clang`` to avoid an import cycle);
+``dumper_clang.py`` imports back what it needs (``_SCOPE_NODE_KINDS`` for its
+own ``_ClangAstParser._walk``, ``_unwrap_expr`` for its own
+``_evaluated_int_value``, ``_field_initializer_value``/``_initializer_value``
+for field/param construction).
+
+Value representation: a lone literal (after stripping wrapper casts) keeps
+its human-readable value (``"42"``); anything compound reduces to a short,
+build-stable structural fingerprint (``"expr:" + sha256(...)[:16]``) so two
+different compound expressions compare unequal while the same one is stable
+across builds — deliberately NOT cross-comparable with castxml's verbatim
+source expression (see ``diff_types``'s/``diff_symbols``'s same-producer
+gates, and ``diff_symbols_param_defaults.py`` for the schema-version gate
+this fingerprint's own algorithm churn needed).
+"""
+
+from __future__ import annotations
+
+import hashlib
+import json
+import re
+from collections.abc import Callable
+from typing import Any
+
+from .diff_cxx_rules import itanium_scope_components
+
+#: Decl contexts we descend into, tracking the enclosing scope name so a
+#: namespace/class-qualified constant key is built (``ns::C::kLimit``).
+#: Shared with ``dumper_clang._ClangAstParser._walk``'s own public-surface
+#: qualified-name building (imported back from there) — kept as ONE
+#: definition rather than two independently-drifting copies.
+_SCOPE_NODE_KINDS = frozenset(
+    {"NamespaceDecl", "CXXRecordDecl", "RecordDecl", "LinkageSpecDecl"}
+)
+#: Literal node kinds whose ``value`` is a stable, human-meaningful constant.
+_LITERAL_NODE_KINDS = frozenset(
+    {
+        "IntegerLiteral",
+        "FloatingLiteral",
+        "CharacterLiteral",
+        "StringLiteral",
+        "CXXBoolLiteralExpr",
+        "FixedPointLiteral",
+    }
+)
+#: Single-child wrapper expressions to descend through before reading a literal.
+_WRAPPER_EXPR_KINDS = frozenset(
+    {
+        "ImplicitCastExpr",
+        "CStyleCastExpr",
+        "CXXStaticCastExpr",
+        "ConstantExpr",
+        "ExprWithCleanups",
+        "ParenExpr",
+        "CXXFunctionalCastExpr",
+        "MaterializeTemporaryExpr",
+    }
+)
+
+#: A lazy, memoized provider for :func:`_index_decl_id_qualified_names`'s
+#: index -- a CALLABLE, not the dict itself, so the (potentially expensive,
+#: whole-AST) index is only ever built if a fingerprint actually needs to
+#: resolve a ``referencedDecl`` (Codex review, fresh evidence, second round:
+#: gating the eager ``self._id_index()`` call on ``hasInClassInitializer``/
+#: ``_param_has_default`` alone still built it for a plain LITERAL default,
+#: e.g. `int timeout = 30;`, which never reaches the point in
+#: :func:`_canonical_expr` that would actually consult it).
+_IdIndexProvider = Callable[[], dict[str, str]]
+
+
+def _unwrap_expr(node: dict[str, Any]) -> dict[str, Any]:
+    """Descend through single-child wrapper expressions (casts, ConstantExpr…)."""
+    cur = node
+    while isinstance(cur, dict) and cur.get("kind") in _WRAPPER_EXPR_KINDS:
+        inner = [c for c in cur.get("inner", []) or [] if isinstance(c, dict)]
+        if len(inner) != 1:
+            break
+        cur = inner[0]
+    return cur
+
+
+def _init_expr(node: dict[str, Any]) -> dict[str, Any] | None:
+    """The initializer expression child of a Var/Field decl, or ``None``."""
+    candidates = [
+        c
+        for c in node.get("inner", []) or []
+        if isinstance(c, dict)
+        and not str(c.get("kind", "")).endswith(("Decl", "Attr", "Comment"))
+    ]
+    return candidates[-1] if candidates else None
+
+
+def _initializer_value(
+    node: dict[str, Any], id_index: _IdIndexProvider | None = None
+) -> str | None:
+    """A stable value string for a variable's initializer, or ``None`` if absent.
+
+    A lone literal (after stripping wrapper casts) keeps its human-readable value
+    (``42``); any compound initializer is reduced to a short deterministic
+    fingerprint so two different compound expressions compare unequal while the
+    same one is stable across builds. Mirrors the castxml ``init`` value as a
+    same-backend comparison key (cross-backend constant *values* are not
+    expected to match — the snapshots are still per-backend parity oracles for
+    presence/scope).
+
+    *id_index* (typically :func:`_index_decl_id_qualified_names` over the
+    whole AST root) resolves a referenced declaration's scope, so ``a::VALUE``
+    and ``b::VALUE`` fingerprint distinctly — see :func:`_canonical_expr`.
+    """
+    init = _init_expr(node)
+    if init is None:
+        return None
+    core = _unwrap_expr(init)
+    if core.get("kind") in _LITERAL_NODE_KINDS and "value" in core:
+        return str(core["value"])
+    return _expr_fingerprint(init, id_index)
+
+
+def _field_initializer_value(
+    field: dict[str, Any], id_index: _IdIndexProvider | None = None
+) -> str | None:
+    """A ``TypeField.default`` value for a ``FieldDecl``, or ``None``.
+
+    G31 Phase C: the last of that phase's fact-completeness list that the
+    direct-clang backend genuinely can close (vptr *placement* it still
+    cannot — clang's plain ``-ast-dump=json`` carries no secondary-vtable
+    offsets without the optional ``ABICHECK_CLANG_LAYOUT_TOOL`` companion).
+
+    Presence is taken from clang's own ``hasInClassInitializer`` flag rather
+    than from "does this decl have a non-attribute ``inner`` child" the way
+    ``_param_has_default`` does, because a ``FieldDecl``'s ``inner`` list
+    is overloaded: a **bitfield width** is nested there as a ``ConstantExpr``
+    too. Verified against real Clang 18 output — ``int bf : 3;`` (no
+    initializer at all) nests exactly one ``ConstantExpr`` child with
+    ``value: "3"``, which the param-style "any non-attribute child" heuristic
+    would read as a default member initializer of ``3``, fabricating an
+    initializer for a field that has none. ``hasInClassInitializer`` is
+    present-only-when-true (absent, never present-and-``false``, for both the
+    plain and the bitfield case), matching the same convention
+    ``scopedEnumTag`` uses.
+
+    The width/initializer ordering is what makes reusing
+    :func:`_initializer_value` safe for the combined form: ``int bfi : 3 = 2;``
+    nests the width ``ConstantExpr`` *first* and the initializer second, and
+    :func:`_init_expr` takes the last non-``Decl``/``Attr``/``Comment`` child
+    — so the initializer wins. (A trailing ``[[deprecated]]`` is a
+    ``DeprecatedAttr``, which that same filter already drops; confirmed
+    against real output, where clang emits it *after* the initializer.)
+
+    Value representation matches :func:`_initializer_value`'s contract, i.e.
+    the same-backend-comparable one ``Param.default`` already uses on this
+    backend: a bare literal keeps its readable value, anything compound
+    reduces to a structural fingerprint. That is deliberately NOT
+    cross-comparable with castxml's verbatim source expression — see
+    ``diff_types._diff_field_default_initializer``'s same-producer gate.
+    """
+    if not field.get("hasInClassInitializer"):
+        return None
+    return _initializer_value(field, id_index)
+
+
+def _expr_fingerprint(
+    node: dict[str, Any], id_index: _IdIndexProvider | None = None
+) -> str:
+    """A short, build-stable structural fingerprint of an expression subtree."""
+    blob = json.dumps(_canonical_expr(node, id_index), sort_keys=True).encode("utf-8")
+    return "expr:" + hashlib.sha256(blob).hexdigest()[:16]
+
+
+def _index_decl_id_qualified_names(root: dict[str, Any]) -> dict[str, str]:
+    """Map every named declaration's clang ``id`` to its scope-qualified name.
+
+    A single, dedicated pass over the WHOLE AST root (independent of
+    ``_ClangAstParser``'s own categorizing walk, which only tracks the
+    ABI-surface kinds it collects): a ``DeclRefExpr``'s ``referencedDecl``
+    can name any declaration in the TU, so this needs to see everything.
+
+    Feeds :func:`_canonical_expr`'s referenced-declaration fingerprinting: a
+    ``referencedDecl`` stub is compact and carries only a bare, unqualified
+    ``name`` (Codex review, fresh evidence: `a::VALUE`/`b::VALUE` share the
+    byte-identical stub). Its own ``id`` IS unique but is a compile-time-only
+    memory address, never stable across builds, so it is exchanged for this
+    index's qualified-name string rather than hashed directly.
+
+    Same namespace/class scope-tracking rule as
+    ``_ClangAstParser._walk``/``_SCOPE_NODE_KINDS``, EXCEPT for
+    ``ClassTemplateSpecializationDecl`` (Codex review, second round):
+    distinct specializations of the same template (``A<int>`` vs.
+    ``A<long>``) both expose only the bare primary-template name ``"A"``, no
+    template-argument spelling at all on the node itself.
+    ``_SCOPE_NODE_KINDS`` deliberately stays untouched (it also drives
+    ``_ClangAstParser._walk``'s own public-surface qualified names, a much
+    larger blast radius); this function special-cases the kind locally via
+    :func:`_specialization_scope_key`.
+
+    A redeclaration (e.g. a function declared then defined) shares its real
+    entity's name, so the first sighting of a given ``id`` is kept rather
+    than overwritten.
+    """
+    index: dict[str, str] = {}
+
+    def walk(node: Any, scope: tuple[str, ...]) -> None:
+        if not isinstance(node, dict):
+            return
+        kind = node.get("kind")
+        name = node.get("name") or ""
+        node_id = node.get("id")
+        if isinstance(node_id, str) and name:
+            index.setdefault(node_id, "::".join((*scope, name)) if scope else name)
+        is_specialization = kind == "ClassTemplateSpecializationDecl"
+        if is_specialization and name:
+            # A representative member's own MANGLED name encodes the
+            # template arguments (`_ZN1AIiE5VALUEE` vs `_ZN1AIlE5VALUEE` for
+            # `VALUE`) and, unlike this node's own `id` (a compile-time
+            # memory address), is build-stable -- so it disambiguates
+            # without ever hashing an unstable value into the persisted
+            # fingerprint _canonical_expr ultimately produces. Falls back to
+            # the bare (collision-prone) name when no direct child carries
+            # one at all -- a rarer, accepted degradation, the same
+            # "conservative fallback over a wrong guess" convention this
+            # whole module already follows elsewhere.
+            disambiguator = _specialization_scope_key(node)
+            scope_name = f"{name}#{disambiguator}" if disambiguator else name
+        else:
+            scope_name = name
+        scope_forming = kind in _SCOPE_NODE_KINDS or is_specialization
+        child_scope = (*scope, scope_name) if scope_forming and scope_name else scope
+        for child in node.get("inner", []) or []:
+            walk(child, child_scope)
+
+    walk(root, ())
+    return index
+
+
+def _specialization_scope_key(node: dict[str, Any]) -> str:
+    """A build-stable, MEMBER-ORDER-INDEPENDENT identity for a
+    ``ClassTemplateSpecializationDecl``, or ``""`` when none can be derived.
+
+    An earlier version used whichever direct child happened to be FIRST with
+    a mangled name -- unstable to unrelated edits (Codex review: inserting
+    `static constexpr int AAA` before an unchanged `VALUE` changed the
+    disambiguator from `VALUE`'s own mangled name to `AAA`'s, perturbing
+    every OTHER declaration referencing `VALUE` though nothing about it
+    changed).
+
+    Fixed via the SCOPE portion of a representative member's mangled name
+    (:func:`diff_cxx_rules.itanium_scope_components`), e.g.
+    ``_ZN1AIiE5VALUEE`` -> ``["AIiE", "VALUE"]``: dropping the trailing leaf
+    leaves ``["AIiE"]``, identical regardless of which member contributed
+    it. Tries every child until one parses -- a special member/operator's
+    name isn't parseable this way, and stopping there would reopen the
+    same instability.
+
+    KNOWN GAP, Itanium-only (Codex review, fresh evidence, third round): a
+    ``clang-cl``/``--target=*-windows-msvc`` snapshot mangles a
+    specialization member as e.g. ``?VALUE@?$A@H@@2HB``, and
+    ``itanium_scope_components`` naturally returns ``None`` for it (wrong
+    scheme). Deliberately NOT falling back to
+    ``diff_cxx_rules.msvc_scope_components`` here -- verified it ALSO
+    returns ``None`` for this exact input, by its own documented design
+    (a component starting with the template marker ``?$`` is rejected
+    outright, and a specialization's class-name component always does) --
+    so unlike the Mach-O/MSVC fallbacks already wired into
+    ``_index_decl_id_qualified_names``'s sibling helpers, adding that call
+    here would be dead code on every real input this function ever sees
+    (it is only invoked for a ``ClassTemplateSpecializationDecl`` member,
+    which is unconditionally template-mangled). ``A<int>``/``A<long>`` on
+    an MSVC-targeted direct-clang snapshot therefore still collide to the
+    bare ``"A"`` -- a real fix needs genuine MSVC template-argument
+    decoding, a new capability with no existing test/caller to verify it
+    against yet (same "a wrong guess here is worse than the pre-existing
+    gap" reasoning AGENTS.md already records for the sibling MSVC
+    toolchain-profile gap).
+    """
+    for child in node.get("inner", []) or []:
+        if not isinstance(child, dict):
+            continue
+        mangled = child.get("mangledName")
+        if not isinstance(mangled, str) or not mangled:
+            continue
+        comps = itanium_scope_components(mangled)
+        if comps and len(comps) > 1:
+            return "::".join(comps[:-1])
+    return ""
+
+
+#: Matches clang's location-bearing anonymous/lambda type spellings, e.g.
+#: ``"(unnamed enum at t.hpp:1:1)"`` or ``"S::(lambda at t.hpp:2:38)"`` --
+#: group 2 (the tag kind, absent for "lambda") is the only part kept.
+_ANON_TYPE_LOCATION_RE = re.compile(r"\((unnamed (\w+)|lambda) at [^)]*\)")
+
+
+def _normalize_qual_type(qual_type: str) -> str:
+    """Strip the source location out of an anonymous-tag/lambda ``qualType``
+    before it's folded into a build-stable fingerprint.
+
+    clang spells an anonymous enum/struct/union/class's type as ``"(unnamed
+    <kind> at <file>:<line>:<col>)"``, and a lambda's closure type as
+    ``"(lambda at ...)"`` -- both embed an absolute path/line (Codex review:
+    two checkout paths, or a blank line before an anonymous type/lambda
+    default, produced different fingerprints for an unchanged initializer).
+    A lambda's spelling recurs throughout any type built on it (e.g. a
+    `std::function<...>`-wrapped lambda's instantiation chain), so every
+    occurrence is substituted, not just the outermost. The "unnamed
+    <kind>"/"lambda" portion is kept; only the location collapses to a
+    fixed placeholder. A named (the common case) qualType passes through
+    unchanged.
+    """
+
+    def _repl(m: re.Match[str]) -> str:
+        kind = m.group(2)
+        return f"(unnamed {kind})" if kind else "(lambda)"
+
+    return _ANON_TYPE_LOCATION_RE.sub(_repl, qual_type)
+
+
+def _canonical_expr(node: Any, id_index: _IdIndexProvider | None = None) -> Any:
+    """Reduce an expression node to a structural form (drop ids/locations)."""
+    if not isinstance(node, dict):
+        return node
+    out: dict[str, Any] = {}
+    for key in ("kind", "value", "opcode", "name", "castKind"):
+        if key in node:
+            out[key] = node[key]
+    type_obj = node.get("type")
+    if isinstance(type_obj, dict) and "qualType" in type_obj:
+        out["type"] = _normalize_qual_type(type_obj["qualType"])
+    # A UnaryExprOrTypeTraitExpr (sizeof/alignof/... applied to a TYPE, not
+    # an expression) stores its operand EXCLUSIVELY in "argType" -- "type" is
+    # just the trait's result type (always "unsigned long" for sizeof,
+    # regardless of operand) (Codex review: `sizeof(int)` vs `sizeof(long
+    # long)` fingerprinted identically without this). Other trait-expression
+    # operand shapes are an unverified, narrower residual gap.
+    arg_type_obj = node.get("argType")
+    if isinstance(arg_type_obj, dict) and "qualType" in arg_type_obj:
+        out["argType"] = _normalize_qual_type(arg_type_obj["qualType"])
+    referenced = node.get("referencedDecl")
+    if isinstance(referenced, dict):
+        # A DeclRefExpr's own top-level keys never identify WHICH declaration
+        # it names -- that lives only in this compact stub, previously
+        # dropped entirely (Codex review: `int x = DEFAULT_A;` vs `int x =
+        # DEFAULT_B;` fingerprinted identically without this -- affects both
+        # TypeField.default and the pre-existing Param.default, which share
+        # this helper). Its own "id" is a compile-time memory address, never
+        # stable across builds, so only "kind"/"name"/"type" are kept.
+        ref_out: dict[str, Any] = {}
+        for key in ("kind", "name"):
+            if key in referenced:
+                ref_out[key] = referenced[key]
+        ref_type = referenced.get("type")
+        if isinstance(ref_type, dict) and "qualType" in ref_type:
+            ref_out["type"] = _normalize_qual_type(ref_type["qualType"])
+        # The bare "name" above collides across scopes (Codex review, second
+        # round): `a::VALUE` vs `b::VALUE` share it. Resolve via id_index
+        # when available -- falls back to bare-name-only (still better than
+        # nothing) when the id isn't found, e.g. a builtin.
+        ref_id = referenced.get("id")
+        if id_index is not None and isinstance(ref_id, str):
+            # Called lazily, right here -- not by the caller -- so a literal
+            # default (the common case, which never reaches this branch at
+            # all) never pays for the index build; only a real
+            # referencedDecl does. The provider itself is expected to be
+            # memoized (_ClangAstParser._id_index()), so repeated resolution
+            # within one parse is still a single AST walk.
+            qualified = id_index().get(ref_id)
+            if qualified is not None:
+                ref_out["qualified_name"] = qualified
+        if ref_out:
+            out["referencedDecl"] = ref_out
+    inner = node.get("inner")
+    if isinstance(inner, list):
+        out["inner"] = [_canonical_expr(c, id_index) for c in inner]
+    return out
